Move my tests to ItemTest, add data provider for them, add testSortByPriceHighToLow
