+++ src/test/java/FirstTest.java
@@ -1,14 +1,12 @@
 import org.openqa.selenium.*;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.Select;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 
@@ -17,19 +15,6 @@ public class FirstTest extends BaseTest {
     private static final String LOGIN_TEST = "Testlogin";
     private static final String PASSWORD_TEST = "Testpassword";
     private static final String EMAIL_TEST = "test@gmail.com";
-    private static final String ITEM_CATEGORY = "jeans";
-
-    private int countItemsContainingItemText(List<WebElement> items) {
-        int count = 0;
-        for (WebElement item : items) {
-            String itemText = item.getText().toLowerCase();
-            if (itemText.contains(ITEM_CATEGORY.toLowerCase())) {
-                count++;
-            }
-        }
-
-        return count;
-    }
 
     @Test
     public void testFirst() throws InterruptedException {
@@ -455,79 +440,6 @@ public class FirstTest extends BaseTest {
 
         Assert.assertEquals("Blue Tshirt", t_shirt);
         Assert.assertEquals(driver.getCurrentUrl(), URL_t_shirt);
-    }
-
-    @Test
-    public void testSearchReturnsAllItemsInCategories() {
-        driver.findElement(By.xpath("//a[@href='/store']")).click();
-        driver.findElement(By.xpath("//input[@type='search']")).sendKeys(ITEM_CATEGORY);
-        driver.findElement(By.xpath("//button[@type='submit']")).click();
-
-        List<WebElement> searchResultList = driver.findElements(By.xpath("//ul//h2"));
-        Assert.assertFalse(searchResultList.isEmpty(), "Search results are empty.");
-        int countItemBySearch = countItemsContainingItemText(searchResultList);
-
-        driver.findElement(By.xpath("//li[@id='menu-item-1228']//a[text()='Men']")).click();
-        List<WebElement> menItemsList = driver.findElements(By.xpath("//ul//h2"));
-        int countItemInMenResult = countItemsContainingItemText(menItemsList);
-
-        driver.findElement(By.xpath("//li[@id='menu-item-1229']//a[text()='Women']")).click();
-        List<WebElement> womenItemsList = driver.findElements(By.xpath("//ul//h2"));
-        int countItemInWomenResult = countItemsContainingItemText(womenItemsList);
-
-        Assert.assertEquals(countItemBySearch, countItemInMenResult + countItemInWomenResult,
-                "Search box did not find all the items with item name '"
-                        + ITEM_CATEGORY + "' or find extra items");
-    }
-
-    @Test
-    public void testVerifyItemsAlphabeticalOrder() {
-        driver.findElement(By.xpath("//li[@id='menu-item-1227']")).click();
-
-        List<String> allItemList = new ArrayList<>();
-        boolean hasNextPage = true;
-
-        while (hasNextPage) {
-            List<WebElement> itemList = driver.findElements(By.xpath("//ul//h2"));
-            for (WebElement item : itemList) {
-                allItemList.add(item.getText());
-            }
-
-            try {
-                WebElement nextPageArrow = driver.findElement(By.xpath("//a[@class='next page-numbers']"));
-                nextPageArrow.click();
-
-            } catch (NoSuchElementException e) {
-                hasNextPage = false;
-            }
-        }
-
-        List<String> alphabeticalAllItemList = new ArrayList<>(allItemList);
-        Collections.sort(alphabeticalAllItemList);
-
-        Assert.assertEquals(allItemList, alphabeticalAllItemList, "Items are not in alphabetical order");
-    }
-
-    @Test
-    public void testSortByPriceLowToHigh() {
-        driver.findElement(By.id("menu-item-1230")).click();
-
-        WebElement dropdown = driver.findElement(By.xpath("//select[@name='orderby']"));
-        Select select = new Select(dropdown);
-        select.selectByVisibleText("Sort by price: low to high");
-
-        List<String> actualPriceList = new ArrayList<>();
-        List<WebElement> priceList = driver.findElements(
-                By.xpath("//span[@class='price']/*[not(@aria-hidden='true')]"));
-        for (WebElement price : priceList) {
-            actualPriceList.add(price.getText());
-        }
-
-        List<String> expectedLowToHighPriceList = new ArrayList<>(actualPriceList);
-        Collections.sort(expectedLowToHighPriceList);
-
-        Assert.assertEquals(actualPriceList, expectedLowToHighPriceList,
-                "Prices are not sorted from high to low as expected");
     }
 
     @Test
